+++ src/main/java/com/chijia/erp/model/dto/PurchaseOrderDTO.java
@@ -1,9 +1,8 @@
 package com.chijia.erp.model.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
-
-import com.chijia.erp.model.dto.SaleOrderDTO.ItemDTO;
 
 import lombok.Data;
 
@@ -12,7 +11,7 @@ public class PurchaseOrderDTO {
 	private Long id;
 	private String purchaseNo;
 	private Long supplierId;
-	private Integer totalAmount;
+	private BigDecimal totalAmount;
 	private LocalDateTime purchaseDate;
 	private String remark;
 	private List<ItemDTO> items;
@@ -22,9 +21,11 @@ public class PurchaseOrderDTO {
 		private Long id;
 		private Long productId;
 		private String productName;
-		private Integer quantity;
-		private Integer unitPrice;
-		private Integer subtotal;
+		private BigDecimal quantity;
+		private BigDecimal unitPrice;
+		private BigDecimal subtotal;
 		
 	}
+
+	
 }
